requirements.txt
tqdm~=4.67.0
aiohttp~=3.11.0
asyncpg~=0.30.0
python-dotenv~=1.0.0
SQLAlchemy~=2.0.0
